board 세션 연결 및 유효성 검사

+++ src/main/java/com/tenco/hobby/repository/model/Board.java
@@ -10,9 +10,7 @@ public class Board {
 	private Long id;	
 	private String title;
 	private String content;
-	//private Long userId;
-	private int userId;
-	
+	private Long userId;	
 	private Long hobbyId;
 	private Timestamp createdAt;
